#1 feat: 파일 분리 및 색상선택 옵션 컴포넌트 선언
